pplx-clone-backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
tavily-python
pydantic
aiofiles